update history, room password

+++ API/ERPSystem.Api/Controllers/MeetingLogController.cs
@@ -4,7 +4,7 @@ using ERPSystem.Common.Infrastructure;
 using ERPSystem.Common.Resources;
 using ERPSystem.DataModel;
 using ERPSystem.DataModel.API;
-using ERPSystem.DataModel.MeetingRoom;
+using ERPSystem.DataModel.MeetingLog;
 using ERPSystem.Service.Handler;
 using ERPSystem.Service.Services;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
@@ -14,16 +14,16 @@ using Microsoft.AspNetCore.Mvc;
 namespace ERPSystem.Api.Controllers;
 
 [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
-public class MeetingRoomController : ControllerBase
+public class MeetingLogController : ControllerBase
 {
     private readonly HttpContext _httpContext;
     private readonly IConfiguration _configuration;
-    private readonly IMeetingRoomService _meetingLogService;
+    private readonly IMeetingLogService _meetingLogService;
     private readonly IMapper _mapper;
 
-    public MeetingRoomController(IHttpContextAccessor httpContextAccessor, IConfiguration configuration,
+    public MeetingLogController(IHttpContextAccessor httpContextAccessor, IConfiguration configuration,
         IJwtHandler jwtHandler,
-        IMeetingRoomService meetingLogService, IMapper mapper)
+        IMeetingLogService meetingLogService, IMapper mapper)
     {
         _httpContext = httpContextAccessor.HttpContext;
         _configuration = configuration;
@@ -34,7 +34,7 @@ public class MeetingRoomController : ControllerBase
     /// <summary>
     /// Get list meeting log
     /// </summary>
-    /// <param name="search"></param>
+    /// <param name="userId"></param>
     /// <param name="meetingRoomIds"></param>
     /// <param name="folderLogs"></param>
     /// <param name="pageNumber"></param>
@@ -44,17 +44,16 @@ public class MeetingRoomController : ControllerBase
     /// <returns></returns>
     [HttpGet]
     [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
-    [Route(Constants.Route.ApiMeetingRooms)]
-    [CheckPermission(PagePermission.ActionName.View + PagePermission.Page.Meeting)]
-    public IActionResult Gets(string search, List<int> meetingRoomIds, List<int> folderLogs, int pageNumber = 1, int pageSize = 10,
-        string sortColumn = "Title",
+    [Route(Constants.Route.ApiMeetingLogs)]
+    public IActionResult Gets(int userId, List<int> meetingRoomIds, List<int> folderLogs, int pageNumber = 1, int pageSize = 10,
+        string sortColumn = "CreatedOn",
         string sortDirection = "asc")
     {
-        var meetingLogs = _meetingLogService.GetPaginated(search, meetingRoomIds, folderLogs, pageNumber, pageSize, sortColumn.MeetingRoomColumn(),
+        var meetingLogs = _meetingLogService.GetPaginated(userId, meetingRoomIds, folderLogs, pageNumber, pageSize, sortColumn.MeetingLogColumn(),
             sortDirection, out var recordsTotal,
             out var recordsFiltered);
 
-        var pagingData = new PagingData<MeetingRoomResponseModel>
+        var pagingData = new PagingData<MeetingLogResponseModel>
         {
             Data = meetingLogs,
             Meta = { RecordsTotal = recordsTotal, RecordsFiltered = recordsFiltered }
@@ -68,9 +67,8 @@ public class MeetingRoomController : ControllerBase
     /// <param name="model"></param>
     /// <returns></returns>
     [HttpPost]
-    [Route(Constants.Route.ApiMeetingRooms)]
-    [CheckPermission(PagePermission.ActionName.Add + PagePermission.Page.Meeting)]
-    public IActionResult Add([FromBody] MeetingRoomModel model)
+    [Route(Constants.Route.ApiMeetingLogs)]
+    public IActionResult Add([FromBody] MeetingLogModel model)
     {
         if (!ModelState.IsValid)
         {
@@ -92,8 +90,7 @@ public class MeetingRoomController : ControllerBase
     /// <param name="ids"></param>
     /// <returns></returns>
     [HttpDelete]
-    [Route(Constants.Route.ApiMeetingRooms)]
-    [CheckPermission(PagePermission.ActionName.Delete + PagePermission.Page.Meeting)]
+    [Route(Constants.Route.ApiMeetingLogs)]
     public IActionResult DeleteMulti(List<int> ids)
     {
         if (ids is not { Count: > 0 })
@@ -101,6 +98,13 @@ public class MeetingRoomController : ControllerBase
             return new ApiErrorResult(StatusCodes.Status400BadRequest, MessageResource.BadRequest);
         }
 
+        // check current user map createdBy.
+        bool isUser = _meetingLogService.CheckCreatedBy(ids, _httpContext.User.GetAccountId());
+        if (!isUser)
+        {
+            return new ApiErrorResult(StatusCodes.Status422UnprocessableEntity,
+                MeetingLogResource.msgCanNotDeleteRecordOfAnotherUser);
+        }
 
         bool result = _meetingLogService.Delete(ids);
         if (!result)
@@ -117,8 +121,7 @@ public class MeetingRoomController : ControllerBase
     /// <param name="id"></param>
     /// <returns></returns>
     [HttpGet]
-    [Route(Constants.Route.ApiMeetingRoomsId)]
-    [CheckPermission(PagePermission.ActionName.View + PagePermission.Page.Meeting)]
+    [Route(Constants.Route.ApiMeetingLogsId)]
     public IActionResult GetById(int id)
     {
         var item = _meetingLogService.GetById(id);
@@ -137,9 +140,8 @@ public class MeetingRoomController : ControllerBase
     /// <param name="model"></param>
     /// <returns></returns>
     [HttpPut]
-    [Route(Constants.Route.ApiMeetingRoomsId)]
-    [CheckPermission(PagePermission.ActionName.Edit + PagePermission.Page.Meeting)]
-    public IActionResult Edit(int id, [FromBody] MeetingRoomModel model)
+    [Route(Constants.Route.ApiMeetingLogsId)]
+    public IActionResult Edit(int id, [FromBody] MeetingLogModel model)
     {
         var item = _meetingLogService.GetById(id);
         if (item == null)
@@ -154,7 +156,12 @@ public class MeetingRoomController : ControllerBase
             return new ValidationFailedResult(ModelState);
         }
 
-   
+        bool isUser = _meetingLogService.CheckCreatedBy(new List<int>{id}, _httpContext.User.GetAccountId());
+        if (!isUser)
+        {
+            return new ApiErrorResult(StatusCodes.Status422UnprocessableEntity,
+                MeetingLogResource.msgCanNotEditRecordOfAnotherUser);
+        }
 
         bool result = _meetingLogService.Update(model);
         if (!result)
@@ -171,8 +178,7 @@ public class MeetingRoomController : ControllerBase
     /// <param name="id"></param>
     /// <returns></returns>
     [HttpDelete]
-    [Route(Constants.Route.ApiMeetingRoomsId)]
-    [CheckPermission(PagePermission.ActionName.Delete + PagePermission.Page.Meeting)]
+    [Route(Constants.Route.ApiMeetingLogsId)]
     public IActionResult Delete(int id)
     {
         var item = _meetingLogService.GetById(id);
@@ -180,7 +186,13 @@ public class MeetingRoomController : ControllerBase
         {
             return new ApiErrorResult(StatusCodes.Status404NotFound, MessageResource.ResourceNotFound);
         }
-     
+        // check current user map createdBy.
+        bool isUser = _meetingLogService.CheckCreatedBy(new List<int>(){id}, _httpContext.User.GetAccountId());
+        if (!isUser)
+        {
+            return new ApiErrorResult(StatusCodes.Status422UnprocessableEntity,
+                MeetingLogResource.msgCanNotDeleteRecordOfAnotherUser);
+        }
         bool result = _meetingLogService.Delete(new List<int>{id});
         if (!result)
         {
